+++ src/controllers/livroController.js
@@ -1,6 +1,6 @@
-import mongoose from 'mongoose';
 import { autor } from '../models/Autor.js';
 import livro from '../models/Livro.js';
+import Erro404 from '../errors/Erro404.js';
 
 class LivroController {
   static async listarLivros(req, res, next) {
@@ -19,7 +19,7 @@ class LivroController {
       if (livroEncontrado) {
         res.status(200).json(livroEncontrado);
       } else {
-        res.status(404).json({ message: 'livro não localizado por este ID' });
+        next(new Erro404('livro não localizado por este ID'));
       }
     } catch (error) {
       next(error);
@@ -31,10 +31,10 @@ class LivroController {
       const novoLivro = req.body;
       const autorEncontrado = await autor.findById(novoLivro.autor);
       if (!autorEncontrado) {
-        throw new mongoose.Error.CastError(
-          'ObjectId',
-          novoLivro.autor,
-          'autor'
+        next(
+          new Erro404(
+            'para criar um livro é necessário o ID de um autor já existente'
+          )
         );
       }
       const livroCompleto = {
@@ -54,6 +54,9 @@ class LivroController {
   static async editarLivroPorId(req, res, next) {
     try {
       const id = req.params.id;
+      if (!id) {
+        next(new Erro404('livro não localizado por este ID'));
+      }
       await livro.findByIdAndUpdate(id, req.body);
       res.status(200).json({ message: 'Livro atualizado!' });
     } catch (error) {
@@ -64,6 +67,9 @@ class LivroController {
   static async deletarLivroPorId(req, res, next) {
     try {
       const id = req.params.id;
+      if (!id) {
+        next(new Erro404('livro não localizado por este ID'));
+      }
       await livro.findByIdAndDelete(id);
       res.status(200).json({ message: 'Livro excluído com sucesso!' });
     } catch (error) {
@@ -74,6 +80,9 @@ class LivroController {
   static async listarLivrosPorEditora(req, res, next) {
     const editora = req.query.editora;
     try {
+      if (!editora) {
+        next(new Erro404('livro não localizado por esta editora'));
+      }
       const livrosPorEditora = await livro.find({ editora: editora });
       res.status(200).json(livrosPorEditora);
     } catch (error) {
